criação da pagina de login

+++ conteudo/deletar.php
@@ -2,6 +2,15 @@
 
     include("conexao.php");
 
+    if (!isset($_SESSION)) {
+        session_start();
+    }
+    
+    if (!isset($_SESSION['id_usuario'])) {
+        echo "<script> alert('Por favor, faça login para acessar esta página.'); location.href='index.php?p=login'; </script>";
+        exit;
+    }
+
     $usu_codigo = intval($_GET['usuario']);
 
     $sql_code = "DELETE FROM usuario WHERE codigo = '$usu_codigo'";
